+++ packages/terra-form-radio/src/RadioField.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import 'terra-base/lib/baseStyles';
-import styles from './RadioField.scss';
+import styles from './RadioField.module.scss';
 
 const cx = classNames.bind(styles);